refactor: move vite-node plugin into plugins/
