+++ packages/vite/src/client.ts
@@ -18,7 +18,7 @@ import { DevStyleSSRPlugin } from './plugins/dev-style-ssr'
 import { RuntimePathsPlugin } from './plugins/runtime-paths'
 import { TypeCheckPlugin } from './plugins/type-check'
 import { ModulePreloadPolyfillPlugin } from './plugins/module-preload-polyfill'
-import { ViteNodePlugin } from './vite-node'
+import { ViteNodePlugin } from './plugins/vite-node'
 import { createViteLogger } from './utils/logger'
 import { StableEntryPlugin } from './plugins/stable-entry'
 import { AnalyzePlugin } from './plugins/analyze'
