+++ packages/vite/src/server.ts
@@ -9,7 +9,7 @@ import type { Nitro } from 'nitropack/types'
 import escapeStringRegexp from 'escape-string-regexp'
 import type { ViteBuildContext } from './vite'
 import { createViteLogger } from './utils/logger'
-import { writeDevServer } from './vite-node'
+import { writeDevServer } from './plugins/vite-node'
 import { writeManifest } from './manifest'
 import { transpile } from './utils/transpile'
 import { SourcemapPreserverPlugin } from './plugins/sourcemap-preserver'
